+++ src/Regressors/DummyRegressor.php
@@ -78,6 +78,14 @@ class DummyRegressor implements Learner, Persistable
      */
     public function predict(Dataset $dataset) : array
     {
-        return array_map([$this->strategy, 'guess'], $dataset->samples());
+        $n = $dataset->numRows();
+
+        $predictions = [];
+
+        for ($i = 0; $i < $n; $i++) {
+            $predictions[] = $this->strategy->guess();
+        }
+
+        return $predictions;
     }
 }