+++ tests/Regressors/GradientBoostTest.php
@@ -8,7 +8,9 @@ use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Unlabeled;
 use Rubix\ML\Regressors\GradientBoost;
+use Rubix\ML\Regressors\DummyRegressor;
 use Rubix\ML\Regressors\RegressionTree;
+use Rubix\ML\Other\Strategies\BlurryMean;
 use Rubix\ML\Datasets\Generators\SwissRoll;
 use PHPUnit\Framework\TestCase;
 use InvalidArgumentException;
@@ -28,7 +30,7 @@ class GradientBoostTest extends TestCase
     {
         $this->generator = new SwissRoll(4., -7., 0., 1., 0.3);
 
-        $this->estimator = new GradientBoost(new RegressionTree(4), 130, 0.1, 0.2, 1e-4, 1e-5);
+        $this->estimator = new GradientBoost(new DummyRegressor(new BlurryMean(0.)), new RegressionTree(3), 30, 0.1, 0.5, 1e-4);
     }
 
     public function test_build_regressor()